I can use u64 - rather than u32 I have a 64 bit machine and it is obtuse to use 32-bit - may as well code all problems in 64-bit

+++ rust/pe001/src/main.rs
@@ -1,6 +1,6 @@
-fn pe001() -> u32 {
-    let imax:u32 = 1000;
-    let mut total:u32 = 0;
+fn pe001() -> u64 {
+    let imax:u64 = 1000;
+    let mut total:u64 = 0;
 
     for i in 0 .. imax {
         if i % 3 == 0 || i % 5 == 0 {
